// types/notication.d.ts
export interface NotificationType {
  company: string;
  companyId: string;
  content: string;
  dateCreated: string;
  dateModified: string;
  emailFrom: string;
  emailTo: string;
  file: string;
  formFIle: string;
  messageId: number;
  phoneNumber: number;
  status: boolean;
  subject: string;
  userCreated: string;
  userModified: string;
}